Print the kmer and its distance from the motif
The total_distance() function takes two arguments: a list of strings motif and a string kmer. It returns the total Hamming distance between the motif and the kmer.

The function first gets the length of the kmer using len(kmer), then initializes a variable distance to 0. It then iterates through each position in the kmer using for i in range(k). At each position, it creates a list column of the characters at that position in each string of the motif list using a list comprehension: [motif[j][i] for j in range(len(motif))]. It then adds the Hamming distance between the character in the kmer at that position and each character in the column list to the distance variable using another list comprehension: sum(1 for c in column if c != kmer[i]). Finally, it returns the distance variable.

The main part of the code defines the motif list and the kmers list. It then iterates through each kmer in the kmers list using for kmer in kmers: and calls the total_distance() function with the motif and the current kmer. It then prints the kmer and the distance using print(kmer, distance).